+++ CHANGELOG.md
@@ -3,6 +3,9 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+### Fixed
+* `colsconfig-dialog-content` component:
+    * Fix column names saving in export excel mode.
 
 ## [2.0.0-beta.4] - 2018-10-30
 ### Fixed